+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - `KeyPath` sort method has been added to Array
 - `UIViewController` `Factory` in XCode Templates has been fixed
+- `NetworkProcessor` has been renamed to `NetworkResponseProcessor`
 
 #### [2.0.2(11)](https://github.com/VakhoKontridze/VCore/releases/download/2.0.2/VCore.xcframework.zip) — *2021 12 29*